+++ ContentPatcher/release-notes.md
@@ -5,6 +5,7 @@
 * Added support for editing map tiles and tile properties.
 * Added support for arithmetic expressions.
 * Added support for multi-key bindings (like `LeftShift + F3`).
+* `EditData` patches now also copy layers and layer properties from the source map (thanks to mouse!).
 * Patches are now always applied in the order listed.
 * Improved logic for matching tilesheets when applying a map patch.
 * Fixed incorrect warning when using `HasWalletItem` token in 1.12.